requirements.txt
streamlit==1.28.0
pydantic==2.5.0
openai==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.1.0
requests==2.31.0
python-pptx==0.6.23
aiofiles==23.2.1
httpx==0.25.2
openai-agents==0.0.17 